Factor out query part into processed transactions.
This commit doesn't change much, just factoring out the struct
Ledger as the report::process() result.

Eventually we'll have a few methods under Ledger struct to support
features.
