+++ core/src/report.rs
@@ -8,14 +8,17 @@ mod error;
 mod eval;
 mod intern;
 mod price_db;
+pub mod query;
+mod transaction;
 
 use std::borrow::Borrow;
 
 pub use balance::Balance;
-pub use book_keeping::{process, Posting, ProcessOptions, Transaction};
+pub use book_keeping::{process, ProcessOptions};
 pub use context::{Account, ReportContext};
 pub use error::ReportError;
 pub use price_db::{load_price_db, PriceDBError};
+pub use transaction::{Posting, Transaction};
 
 use crate::{load, syntax::plain::LedgerEntry};
